+++ src/deps.cljs
@@ -1,7 +1,4 @@
-{:npm-deps     {"react"              "16.9.0"
-                "react-dom"          "16.9.0"}
-
- :npm-dev-deps {"shadow-cljs"           "2.8.83"
+{:npm-dev-deps {"shadow-cljs"           "2.8.83"
                 "karma"                 "4.4.1"
                 "karma-chrome-launcher" "3.1.0"
                 "karma-cljs-test"       "0.1.0"